+++ routes/Register/registerRoute.js
@@ -11,20 +11,24 @@ bcrypt.hash = util.promisify(bcrypt.hash);
 // to send SMS
  const Nexmo = require('nexmo');
 
- /*const nexmo = new Nexmo({
+ const nexmo = new Nexmo({
    apiKey: process.env.APIKEY,
    apiSecret: process.env.APISECRET,
  });
-*/
- let patient = require('../../models/patient');
- let paramedic =require('../../models/paramedic');
 
+const patient = require('../../models/patient');
+const paramedic =require('../../models/paramedic');
+const doctor =require('../../models/doctor'); 
  async function register(req,res,type){
     const name = req.body.name;
     const phoneNo = req.body.phoneNo;
     const birthDate = req.body.birthDate;
     const gender = req.body.gender;
     const password = req.body.password;
+    // For paramedics and doctors
+    const rating  = req.body.rating;
+    // For doctors
+    const specialization = req.body.specialization;
 
     //check if the number is valid
     if (!phoneNo || isNaN(+phoneNo) || phoneNo.length !== 12 || phoneNo.substring(0, 3) !== "201") {
@@ -46,10 +50,12 @@ bcrypt.hash = util.promisify(bcrypt.hash);
         gender,
         password,
         _id:phoneNo,
-        randomCode
+        randomCode,
+        rating,
+        specialization
     };
     newAccount.password = await hashPasswords(newAccount);
-    console.log(newAccount)
+    console.log(newAccount);
     const result = await client.get(phoneNo);
 	if(result){
 		return res.status(409).json({"Error":"Account is created and needs confirmation"});
@@ -61,7 +67,7 @@ bcrypt.hash = util.promisify(bcrypt.hash);
     const from = 'Server';
     const to = newAccount._id;
     const text = `Code for verification is : ${randomCode}`;
-    //nexmo.message.sendSms(from, to, text);
+    nexmo.message.sendSms(from, to, text);
 	return res.status(201).json("created successfully");
 }
 
@@ -93,18 +99,15 @@ router.post('/confirmation', async (req, res) => {
         let account;
         if(!result.rating){
             account = new patient(result);
+            dbManger.addPatient(account);
         } else if(!result.specialization){
             account = new paramedic(result);
+            dbManger.addParamedic(account);
         } else {
-            // doctor
+            account = new doctor(result);
+            dbManger.addParamedic(account);
         }
-        client.del(phoneNo);
-        // More logic to be added 
-        if(result.hasOwnProperty('_id')){
-            if(!dbManger.addPatient(account)){
-                return res.status(202).json({"Error":"Try again later"})
-            }
-        }
+        client.hdel(phoneNo,"confirmation");
         return res.status(201).json("Account activated");
     } else {
         return res.status(404).json({"Error":"Not found"});
@@ -119,9 +122,9 @@ router.put('/resend_code', async (req, res) => {
         const randomCode = Math.random().toString().substring(2,6);
         result.randomCode = randomCode;
         const from = 'Server';
-        const to = `2 +${newAccount._id}`;
+        const to = newAccount._id;
         const text = `Code for verification is : ${randomCode}`;
-        //nexmo.message.sendSms(from, to, text);
+        nexmo.message.sendSms(from, to, text);
         client.set(phoneNo,result,"EX",60*60);
         return res.status(201).json("New confirmation code sent");
     } else {
@@ -130,13 +133,17 @@ router.put('/resend_code', async (req, res) => {
 });
 
 async function numberPredefined(phoneNo){
-    let flag = false;
-    await patient.find({_id:phoneNo},(err,res)=>{
-        if(res.length !== 0){
-            flag = true;
-        }
-    });
-    return flag;
+    let result = await dbManger.findParamedic({_id:phoneNo});
+    if(result){ 
+        return true;
+    }
+    result = await dbManger.findPatient({_id:phoneNo});
+    if(result)
+        return true;
+    result = await dbManger.findDoctor({_id:phoneNo});
+    if(result)
+        return true;
+    return false;
 }
 
 router.get('/welcome/info',async(req,res)=>{
